fix: failing unit test
