+++ ui/apps/akasha/src/components/__tests__/feed-page.spec.tsx
@@ -1,6 +1,8 @@
 import * as React from 'react';
 import FeedPage from '../feed-page/feed-page';
+import * as extension from '../extension';
 
+import { InlineEditor } from '../../extensions/inline-editor/inline-editor';
 import {
   screen,
   renderWithAllProviders,
@@ -9,6 +11,7 @@ import {
 } from '@akashaorg/af-testing';
 import { AnalyticsProvider } from '@akashaorg/ui-awf-hooks/lib/use-analytics';
 import { act } from 'react-dom/test-utils';
+import * as hooks from '@akashaorg/ui-awf-hooks/lib/use-profile';
 
 describe('< FeedPage /> component', () => {
   const BaseComponent = ({ loginState }) => (
@@ -16,6 +19,20 @@ describe('< FeedPage /> component', () => {
       <FeedPage {...genAppProps()} showLoginModal={jest.fn()} loginState={loginState} />
     </AnalyticsProvider>
   );
+
+  beforeAll(() => {
+    jest
+      .spyOn(extension, 'Extension')
+      .mockReturnValue(
+        <InlineEditor {...genAppProps()} extensionData={{ name: 'post', action: 'post' }} />,
+      );
+
+    (
+      jest.spyOn(hooks, 'useGetProfile') as unknown as jest.SpyInstance<{
+        status: string;
+      }>
+    ).mockReturnValue({ status: 'success' });
+  });
 
   it('should render feed page for anonymous users', async () => {
     await act(async () => {
@@ -31,7 +48,6 @@ describe('< FeedPage /> component', () => {
     await act(async () => {
       renderWithAllProviders(<BaseComponent loginState={genLoggedInState(true)} />, {});
     });
-
     expect(screen.getAllByTestId('avatar-image')).not.toBeNull();
     expect(screen.getByText(/Share your thoughts/i)).toBeInTheDocument();
   });
