+++ ui/apps/akasha/src/extensions/inline-editor/index.tsx
@@ -1,0 +1,43 @@
+import * as React from 'react';
+import singleSpaReact from 'single-spa-react';
+import ReactDOM from 'react-dom';
+import DS from '@akashaorg/design-system';
+
+import { RootExtensionProps } from '@akashaorg/typings/ui';
+import { withProviders, ThemeWrapper } from '@akashaorg/ui-awf-hooks';
+import { I18nextProvider } from 'react-i18next';
+import { InlineEditor } from './inline-editor';
+
+const { ErrorLoader } = DS;
+
+const Wrapped = (props: RootExtensionProps) => {
+  return (
+    <I18nextProvider i18n={props.plugins['@akashaorg/app-translation']?.translation?.i18n}>
+      <InlineEditor {...props} />
+    </I18nextProvider>
+  );
+};
+
+const reactLifecycles = singleSpaReact({
+  React,
+  ReactDOM,
+  rootComponent: withProviders(Wrapped),
+  renderType: 'createRoot',
+  errorBoundary: (err, errorInfo, props: RootExtensionProps) => {
+    if (props.logger) {
+      props.logger.error(`${JSON.stringify(errorInfo)}, ${errorInfo}`);
+    }
+
+    return (
+      <ThemeWrapper {...props}>
+        <ErrorLoader type="script-error" title="Error in editor modal" details={err.message} />
+      </ThemeWrapper>
+    );
+  },
+});
+
+export const bootstrap = reactLifecycles.bootstrap;
+
+export const mount = reactLifecycles.mount;
+
+export const unmount = reactLifecycles.unmount;
