+++ ui/apps/akasha/src/extensions/inline-editor/inline-editor.tsx
@@ -1,12 +1,9 @@
 import * as React from 'react';
-import singleSpaReact from 'single-spa-react';
-import ReactDOM from 'react-dom';
 import { RootExtensionProps, IPublishData, AnalyticsCategories } from '@akashaorg/typings/ui';
 import DS from '@akashaorg/design-system';
 import {
   uploadMediaToTextile,
   getLinkPreview,
-  withProviders,
   useCreatePost,
   useEditPost,
   usePost,
@@ -16,14 +13,13 @@ import {
   useGetProfile,
   useGetLogin,
   useAnalytics,
-  ThemeWrapper,
   useCreateComment,
 } from '@akashaorg/ui-awf-hooks';
-import { I18nextProvider, useTranslation } from 'react-i18next';
+import { useTranslation } from 'react-i18next';
 
-const { CommentEditor, EntryCardLoading, ErrorLoader } = DS;
+const { CommentEditor, EntryCardLoading } = DS;
 
-const InlineEditorContainer = (props: RootExtensionProps) => {
+export const InlineEditor = (props: RootExtensionProps) => {
   const { t } = useTranslation('app-akasha-integration');
 
   const loginQuery = useGetLogin();
@@ -229,35 +225,3 @@ const InlineEditorContainer = (props: RootExtensionProps) => {
     </>
   );
 };
-
-const Wrapped = (props: RootExtensionProps) => {
-  return (
-    <I18nextProvider i18n={props.plugins['@akashaorg/app-translation']?.translation?.i18n}>
-      <InlineEditorContainer {...props} />
-    </I18nextProvider>
-  );
-};
-
-const reactLifecycles = singleSpaReact({
-  React,
-  ReactDOM,
-  rootComponent: withProviders(Wrapped),
-  renderType: 'createRoot',
-  errorBoundary: (err, errorInfo, props: RootExtensionProps) => {
-    if (props.logger) {
-      props.logger.error(`${JSON.stringify(errorInfo)}, ${errorInfo}`);
-    }
-
-    return (
-      <ThemeWrapper {...props}>
-        <ErrorLoader type="script-error" title="Error in editor modal" details={err.message} />
-      </ThemeWrapper>
-    );
-  },
-});
-
-export const bootstrap = reactLifecycles.bootstrap;
-
-export const mount = reactLifecycles.mount;
-
-export const unmount = reactLifecycles.unmount;
